requirements.txt
streamlit
transformers<5.0.0
torch
Pillow
gTTS
accelerate
bitsandbytes